+++ src/__swaps__/screens/Swap/resources/_selectors/search.ts
@@ -1,4 +1,4 @@
-import { TokenSearchResult } from '../search';
+import { TokenSearchResult } from '../search/search';
 
 export function filterNonTokenIconAssets(tokens: TokenSearchResult) {
   return tokens.filter(token => token.icon_url);
