+++ src/languages/en_US.json
@@ -2059,6 +2059,7 @@
     "token_search": {
       "section_header": {
         "recent": "Recent",
+        "popular": "Popular in Rainbow",
         "favorites": "Favorites",
         "bridge": "Bridge",
         "verified": "Verified",
